Move integration test from snippet to the example bundle
